shooter is locked into repairing and more things abiding by MVC a little more repairing is now an action that is initiated and has to complete fixed that you can hold spacebar instead of spamming

+++ src/main/java/my_project/model/GUI/HPBar.java
@@ -6,7 +6,7 @@ import my_project.model.UI;
 
 public class HPBar extends UI {
 
-    private double health;
+    private int health;
     private CollidableEnvironment environment;
 
     /**
